fetching, adding, removing and updating shifts in the cloud now works for a set user

+++ src/src/Logger/Edit shift/Editing.swift
@@ -21,29 +21,47 @@ class Editing: UIViewController {
     }
 
     @IBAction func save(_ sender: UIButton) {
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
-        
-        let shiftToUpdate = LocalStorage.organizedValues[shiftToEdit[0]][shiftToEdit[1]]
-        context.delete(shiftToUpdate)
-        
-        let newShift = ShiftModel(
-            date: currentTempShift.date,
-            endingTime: currentTempShift.endingTime,
-            startingTime: currentTempShift.startingTime,
-            lunchTime: currentTempShift.lunchTime,
-            note: currentTempShift.note,
-            newPeriod: currentTempShift.newPeriod
-        ).toCoreData()
-        
-        context.insert(newShift)
-        
-        do {
-            try context.save()
-        } catch {
-            print(error)
+        if usingLocalStorage {
+            let appDelegate = UIApplication.shared.delegate as! AppDelegate
+            let context = appDelegate.persistentContainer.viewContext
+            
+            let shiftToUpdate = LocalStorage.organizedValues[shiftToEdit[0]][shiftToEdit[1]]
+            context.delete(shiftToUpdate)
+            
+            let newShift = ShiftModel(
+                date: currentTempShift.date,
+                endingTime: currentTempShift.endingTime,
+                startingTime: currentTempShift.startingTime,
+                lunchTime: currentTempShift.lunchTime,
+                note: currentTempShift.note,
+                newPeriod: currentTempShift.newPeriod
+                ).toCoreData()
+            
+            context.insert(newShift)
+            
+            do {
+                try context.save()
+            } catch {
+                print(error)
+            }
+        } else {
+            let shiftToUpdate = shifts[shiftToEdit[0]][shiftToEdit[1]]
+            
+            let newShift = ShiftModel(
+                date: currentTempShift.date,
+                endingTime: currentTempShift.endingTime,
+                startingTime: currentTempShift.startingTime,
+                lunchTime: currentTempShift.lunchTime,
+                note: currentTempShift.note,
+                newPeriod: currentTempShift.newPeriod
+            )
+            newShift.ID = shiftToUpdate.ID
+            shifts[shiftToEdit[0]][shiftToEdit[1]] = newShift
+            CloudStorage.updateShift(from: shiftToUpdate, with: newShift, user: userId)
         }
-        shouldFetchAllData = true
+        
+        
+        //shouldFetchAllData = true
         performSegue(withIdentifier: "goback", sender: self)
     }
 }
